pyproject.toml
[project]
name = "courses"
version = "0.1.0"
description = ""
authors = [
    {name = "volko",email = "Toxa.Vol@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
flake8 = "^7.1.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
ipython = "^9.0.2"

